requirements.txt
# Dependencias principales del proyecto
chromadb==1.0.9
python-dotenv==1.1.0
boto3==1.38.18
tiktoken==0.9.0

# LangChain - Solo las dependencias que realmente se usan
langchain-core==0.3.60
langchain-aws==0.2.23
langgraph>=0.0.33
